project/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.1
idna==2.9
Mako==1.1.3
MarkupSafe==1.1.1
mongoengine==0.20.0
PyJWT==1.7.1
pymongo==3.10.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
rq==1.4.2
urllib3==1.25.9
